[FIX] moving code to product_pricelist file
